bootstrap: widen OutcomeId to u16 + retry HTTP 408 in fetcher (closes #159)
The May-13 backfill run aborted with 5 wallet failures: 3 Polymarket
multi-outcome markets emitting `outcomeIndex: 999` overflowing every
u8-typed waypoint, and 2 HTTP 408 Request Timeout responses being
routed to SourceError::Fatal instead of being retried.

Production-side changes are width-localised behind two new conversion
impls on OutcomeId (`TryFrom<i64>` and `From<OutcomeId> for i64`) so
any future widening collapses to a single edit in `ids.rs`. The
`MarketResolution.winning_outcome_id` field is promoted from raw `u8`
to `OutcomeId` to kill the `.0` deref footgun at simulation.rs:421.
The fetcher's status-code dispatch is refactored behind a named
`is_retryable_status` helper so adding future codes (425, etc.) is
one word.

No schema migration; SQLite INTEGER already stores any i64. The
`TradeFill.outcome_id` NDJSON-serialisation surface widens to `u16`
(not OutcomeId) — same judgment applied to `resolution_record`.

Tests cover the bug surfaces end-to-end: OutcomeId(999) JSON
round-trip, cache insert/load with winner=999 and u16::MAX,
is_retryable_status policy, scenario_retry_on_408 +
scenario_exhaust_retries_408 against axum mock, and an
end-to-end scenario_outcome_id_widening covering cache idempotency
and resolution-match equality for the multi-outcome path.

+++ crates/backtest/src/report.rs
@@ -190,7 +190,7 @@ pub struct TradeFill {
     pub leader_wallet: String,
     pub operator_id: Option<String>,
     pub market_id: String,
-    pub outcome_id: u8,
+    pub outcome_id: u16,
     pub side: String,
     pub contracts: u64,
     pub signal_price: Decimal,
